new changes

+++ src/Components/Booking/Booking.jsx
@@ -1,24 +1,31 @@
 import React, { useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
+import { useDispatch } from "react-redux";
 import { addSeatsToBooking } from "../../redux/reducers/bookingSlice";
 import FlightSeat from "../FlightSeat/FlightSeat";
 import Swal from "sweetalert2";
 import Meal from "../Meal/Meal";
 import Luggage from "../Luggage/Luggage";
+import useLocalStorage from "../../hooks/useLocalStorage";
 
 const Booking = () => {
+  const dispatch = useDispatch();
   const [selectedSeats, setSelectedSeats] = useState([]);
   const [showMeal, setShowMeal] = useState(false);
   const [showSeat, setShowSeat] = useState(true);
   const [showLuggage, setShowLuggage] = useState(false);
+  const [bookSeat, setBookSeat] = useLocalStorage(
+    "bookSeat",
+    JSON.parse(localStorage.getItem("bookSeat")) || []
+  );
 
-  const bookings = useSelector((state) => state.bookings?.bookings[0]);
-
-  const dispatch = useDispatch();
   const storedBooking = JSON.parse(localStorage.getItem("booking"));
 
+  const currFlightBookedSeat = bookSeat.filter(
+    (bookSeat) => bookSeat.id === storedBooking.id
+  );
+
   const [disableBookings, setDisabledBookings] = useState(
-    bookings?.seats?.booked || []
+    currFlightBookedSeat[0]?.booked || []
   );
 
   const handleSeatSelection = (seatId) => {
@@ -30,23 +37,53 @@ const Booking = () => {
   };
 
   const handleSubmit = () => {
-    selectedSeats &&
-      dispatch(
-        addSeatsToBooking({
-          bookingId: storedBooking.bookingId,
-          seats: { booked: selectedSeats },
-        })
-      );
-    Swal.fire({
-      title: "Success!",
-      text: "Seats Added Successfully",
-      icon: "success",
-      confirmButtonText: "OK",
-    }).then(() => {
-      setDisabledBookings([...disableBookings, ...selectedSeats]);
-      storedBooking.seats.booked = selectedSeats;
-      localStorage.setItem("booking", JSON.stringify(storedBooking));
-    });
+    if (selectedSeats.length > 0) {
+      selectedSeats &&
+        dispatch(
+          addSeatsToBooking({
+            bookingId: storedBooking.bookingId,
+            seats: { booked: selectedSeats },
+          })
+        );
+      Swal.fire({
+        title: "Success!",
+        text: "Seats Added Successfully",
+        icon: "success",
+        confirmButtonText: "OK",
+      }).then(() => {
+        setDisabledBookings([...disableBookings, ...selectedSeats]);
+        storedBooking.seats.booked = selectedSeats;
+        localStorage.setItem("booking", JSON.stringify(storedBooking));
+        const updatedItem = {
+          id: storedBooking.id,
+          booked: [...currFlightBookedSeat[0].booked, ...selectedSeats],
+        };
+
+        const itemIndex = bookSeat.findIndex(
+          (item) => item.id === updatedItem.id
+        );
+
+        const updatedBookSeat = [...bookSeat];
+
+        if (itemIndex !== -1) {
+          updatedBookSeat[itemIndex] = updatedItem;
+        } else {
+          updatedBookSeat.push(updatedItem);
+        }
+
+        setBookSeat(updatedBookSeat);
+        setShowMeal(true);
+        setShowLuggage(false);
+        setShowSeat(false);
+      });
+    } else {
+      Swal.fire({
+        title: "Info!",
+        text: "Please select some seats",
+        icon: "info",
+        confirmButtonText: "OK",
+      });
+    }
   };
 
   const rows = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
@@ -82,7 +119,7 @@ const Booking = () => {
             })}
           </ol>
           <div className="screen-side pt-5">
-            <div className="screen">Screen</div>
+            <div className="screen">Cockpit</div>
           </div>
           <div className="flex items-center justify-center gap-10">
             <button
@@ -94,7 +131,7 @@ const Booking = () => {
             >
               Add Seats
             </button>
-            <button
+            {/* <button
               onClick={() => {
                 setShowMeal(true);
                 setShowLuggage(false);
@@ -106,7 +143,7 @@ const Booking = () => {
               type="submit"
             >
               <i className="fa fa-solid fa-forward"></i>
-            </button>
+            </button> */}
           </div>
         </div>
       )}
